requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fpdf2>=2.7.0
python-docx>=1.1.0
